MOS-12533 TSPID generation methods added
Former-commit-id: ce259fa19e3223efad82caf7990262c2e4ee9012
Former-commit-id: ce13e6bd02002aab167c40f4ba4f5e4dd965266c

+++ kernel/kernel-idgenerator-tspid/src/main/java/io/mosip/kernel/idgenerator/tsp/entity/Tsp.java
@@ -26,7 +26,7 @@ public class Tsp {
 	 */
 	@Id
 	@Column(name = "tsp_id", nullable = false)
-	private long tspId;
+	private int tspId;
 
 	/**
 	 * Created time for TSPID.
